refactor: centralize error handling and conversions across runtime; streamline logic and improve consistency
