+++ pkg/runtime/binary.go
@@ -22,6 +22,20 @@ func NewBinaryFrom(stream io.Reader) (Binary, error) {
 	return values, nil
 }
 
+// ToBinary attempts to convert an arbitrary Value into a Binary type.
+// If the input is already a Binary, it returns it directly.
+// For other types, it converts the input to its string representation and then to a byte slice to create a new Binary.
+// This allows for flexible conversion of various Value types into a Binary format, using their string representation as the basis for the binary data.
+func ToBinary(input Value) Binary {
+	bin, ok := input.(Binary)
+
+	if ok {
+		return bin
+	}
+
+	return NewBinary([]byte(input.String()))
+}
+
 func (b Binary) Type() Type {
 	return TypeBinary
 }
