+++ pkg/runtime/date_time.go
@@ -24,11 +24,11 @@ func NewDateTime(time time.Time) DateTime {
 	return DateTime{time}
 }
 
-func ParseDateTime(input interface{}) (DateTime, error) {
+func ParseDateTime(input any) (DateTime, error) {
 	return ParseDateTimeWith(input, DefaultTimeLayout)
 }
 
-func ParseDateTimeWith(input interface{}, layout string) (DateTime, error) {
+func ParseDateTimeWith(input any, layout string) (DateTime, error) {
 	switch value := input.(type) {
 	case string:
 		t, err := time.Parse(layout, value)
@@ -83,7 +83,7 @@ func ToDateTime(_ context.Context, input Value) (DateTime, error) {
 	}
 }
 
-func MustParseDateTime(input interface{}) DateTime {
+func MustParseDateTime(input any) DateTime {
 	dt, err := ParseDateTime(input)
 
 	if err != nil {
