+++ pkg/runtime/float.go
@@ -1,9 +1,12 @@
 package runtime
 
 import (
+	"context"
 	"encoding/binary"
+	"errors"
 	"fmt"
 	"hash/fnv"
+	"io"
 	"math"
 	"strconv"
 )
@@ -22,7 +25,7 @@ func NewFloat(input float64) Float {
 	return Float(input)
 }
 
-func ParseFloat(input interface{}) (Float, error) {
+func ParseFloat(input any) (Float, error) {
 	if IsNil(input) {
 		return ZeroFloat, nil
 	}
@@ -55,7 +58,7 @@ func ParseFloat(input interface{}) (Float, error) {
 	return ZeroFloat, Error(ErrInvalidType, fmt.Sprintf("expected %s", TypeFloat))
 }
 
-func MustParseFloat(input interface{}) Float {
+func MustParseFloat(input any) Float {
 	res, err := ParseFloat(input)
 
 	if err != nil {
@@ -63,6 +66,72 @@ func MustParseFloat(input interface{}) Float {
 	}
 
 	return res
+}
+
+func ToFloat(ctx context.Context, input Value) (Float, error) {
+	switch val := input.(type) {
+	case Float:
+		return val, nil
+	case Int:
+		return Float(val), nil
+	case String:
+		i, err := strconv.ParseFloat(string(val), 64)
+
+		if err != nil {
+			return ZeroFloat, newConversionError(TypeFloat, err)
+		}
+
+		return Float(i), nil
+	case Boolean:
+		if val {
+			return Float(1), nil
+		}
+
+		return Float(0), nil
+	case DateTime:
+		dt := input.(DateTime)
+
+		if dt.IsZero() {
+			return ZeroFloat, nil
+		}
+
+		return NewFloat(float64(dt.Unix())), nil
+	case List:
+		iterator, err := val.Iterate(ctx)
+
+		if err != nil {
+			return ZeroFloat, err
+		}
+
+		res := ZeroFloat
+
+		for {
+			val, _, err := iterator.Next(ctx)
+			if errors.Is(err, io.EOF) {
+				break
+			}
+
+			if errors.Is(err, ErrTimeout) {
+				break
+			}
+
+			if err != nil {
+				continue
+			}
+
+			f, err := ToFloat(ctx, val)
+
+			if err != nil {
+				continue
+			}
+
+			res += f
+		}
+
+		return res, nil
+	default:
+		return ZeroFloat, newConversionError(TypeFloat, TypeErrorOf(input, TypeFloat))
+	}
 }
 
 func IsNaN(input Float) Boolean {
