+++ pkg/runtime/value_helpers.go
@@ -2,15 +2,11 @@ package runtime
 
 import (
 	"context"
-	"encoding/binary"
 	"errors"
-	"hash/fnv"
 	"io"
 	"math"
 	"math/rand"
 	"reflect"
-	"sort"
-	"strconv"
 	"time"
 )
 
@@ -388,275 +384,4 @@ func ToMap(ctx context.Context, input Value) (Map, error) {
 	default:
 		return NewObject(), nil
 	}
-}
-
-func ToBoolean(input Value) Boolean {
-	if input == None {
-		return False
-	}
-
-	switch val := input.(type) {
-	case Boolean:
-		return val
-	case String:
-		return val != ""
-	case Int:
-		return val != 0
-	case Float:
-		return val != 0
-	case Duration:
-		return val != 0
-	case DateTime:
-		return Boolean(!val.IsZero())
-	default:
-		return True
-	}
-}
-
-func ToFloat(ctx context.Context, input Value) (Float, error) {
-	switch val := input.(type) {
-	case Float:
-		return val, nil
-	case Int:
-		return Float(val), nil
-	case String:
-		i, err := strconv.ParseFloat(string(val), 64)
-
-		if err != nil {
-			return ZeroFloat, newConversionError(TypeFloat, err)
-		}
-
-		return Float(i), nil
-	case Boolean:
-		if val {
-			return Float(1), nil
-		}
-
-		return Float(0), nil
-	case DateTime:
-		dt := input.(DateTime)
-
-		if dt.IsZero() {
-			return ZeroFloat, nil
-		}
-
-		return NewFloat(float64(dt.Unix())), nil
-	case List:
-		iterator, err := val.Iterate(ctx)
-
-		if err != nil {
-			return ZeroFloat, err
-		}
-
-		res := ZeroFloat
-
-		for {
-			val, _, err := iterator.Next(ctx)
-			if errors.Is(err, io.EOF) {
-				break
-			}
-
-			if errors.Is(err, ErrTimeout) {
-				break
-			}
-
-			if err != nil {
-				continue
-			}
-
-			f, err := ToFloat(ctx, val)
-
-			if err != nil {
-				continue
-			}
-
-			res += f
-		}
-
-		return res, nil
-	default:
-		return ZeroFloat, newConversionError(TypeFloat, TypeErrorOf(input, TypeFloat))
-	}
-}
-
-func ToString(input Value) String {
-	switch val := input.(type) {
-	case String:
-		return val
-	default:
-		return NewString(val.String())
-	}
-}
-
-func ToInt(ctx context.Context, input Value) (Int, error) {
-	switch val := input.(type) {
-	case Int:
-		return val, nil
-	case Float:
-		return Int(val), nil
-	case String:
-		i, err := strconv.ParseInt(string(val), 10, 64)
-
-		if err != nil {
-			return ZeroInt, newConversionError(TypeInt, err)
-		}
-
-		return Int(i), nil
-	case Boolean:
-		if val {
-			return Int(1), nil
-		}
-
-		return Int(0), nil
-	case DateTime:
-		dt := input.(DateTime)
-
-		if dt.IsZero() {
-			return ZeroInt, nil
-		}
-
-		return NewInt(int(dt.Unix())), nil
-	case List:
-		iterator, err := val.Iterate(ctx)
-
-		if err != nil {
-			return ZeroInt, err
-		}
-
-		res := ZeroInt
-
-		for {
-			item, _, err := iterator.Next(ctx)
-			if errors.Is(err, io.EOF) {
-				break
-			}
-
-			if errors.Is(err, ErrTimeout) {
-				break
-			}
-
-			if err != nil {
-				continue
-			}
-
-			i, err := ToInt(ctx, item)
-
-			if err != nil {
-				continue
-			}
-
-			res += i
-		}
-
-		return res, nil
-	default:
-		return ZeroInt, newConversionError(TypeInt, TypeErrorOf(input, TypeInt))
-	}
-}
-
-func ToNumber(ctx context.Context, input Value) (Value, error) {
-	switch value := input.(type) {
-	case Int, Float:
-		return value, nil
-	default:
-		return ToFloat(ctx, input)
-	}
-}
-
-func ToIntSafe(ctx context.Context, input Value) Int {
-	result, err := ToInt(ctx, input)
-
-	if err != nil {
-		return ZeroInt
-	}
-
-	if result > 0 {
-		return result
-	}
-
-	return ZeroInt
-}
-
-// ToIntDefault attempts to convert an arbitrary Value into an Int.
-// If the conversion fails or if the resulting Int is not greater than zero, it returns the provided defaultValue.
-// This function is useful for safely converting values to Int while providing a fallback option in case of errors or non-positive results.
-func ToIntDefault(ctx context.Context, input Value, defaultValue Int) (Int, error) {
-	result, err := ToInt(ctx, input)
-
-	if err != nil {
-		return defaultValue, err
-	}
-
-	if result > 0 {
-		return result, nil
-	}
-
-	return defaultValue, nil
-}
-
-// ToBinary attempts to convert an arbitrary Value into a Binary type.
-// If the input is already a Binary, it returns it directly.
-// For other types, it converts the input to its string representation and then to a byte slice to create a new Binary.
-// This allows for flexible conversion of various Value types into a Binary format, using their string representation as the basis for the binary data.
-func ToBinary(input Value) Binary {
-	bin, ok := input.(Binary)
-
-	if ok {
-		return bin
-	}
-
-	return NewBinary([]byte(input.String()))
-}
-
-// Hash computes a hash value for the given typename and content using the FNV-1a hashing algorithm.
-// It concatenates the typename, a colon, and the content bytes to generate a unique hash value.
-// This function can be used to create consistent hash values for different types of content based on their type and content.
-func Hash(typename string, content []byte) uint64 {
-	h := fnv.New64a()
-
-	h.Write([]byte(typename))
-	h.Write([]byte(":"))
-	h.Write(content)
-
-	return h.Sum64()
-}
-
-// MapHash computes a hash value for a map of string keys to Value values.
-// It uses the FNV-1a hashing algorithm to generate a consistent hash based on the keys and their corresponding value hashes.
-// The keys are sorted to ensure that the order of key-value pairs does not affect the resulting hash.
-func MapHash(input map[string]Value) uint64 {
-	h := fnv.New64a()
-
-	keys := make([]string, 0, len(input))
-
-	for key := range input {
-		keys = append(keys, key)
-	}
-
-	// order does not really matter
-	// but it will give us a consistent hash sum
-	sort.Strings(keys)
-	endIndex := len(keys) - 1
-
-	h.Write([]byte("{"))
-
-	for idx, key := range keys {
-		h.Write([]byte(key))
-		h.Write([]byte(":"))
-
-		el := input[key]
-
-		bytes := make([]byte, 8)
-		binary.LittleEndian.PutUint64(bytes, el.Hash())
-
-		h.Write(bytes)
-
-		if idx != endIndex {
-			h.Write([]byte(","))
-		}
-	}
-
-	h.Write([]byte("}"))
-
-	return h.Sum64()
 }
