--- pkg/runtime/zero_divisor_error.go
@@ -1,11 +1,0 @@
-package runtime
-
-import "fmt"
-
-func divisionByZeroError() error {
-	return fmt.Errorf("%w: %w", ErrInvalidOperation, ErrDivisionByZero)
-}
-
-func moduloByZeroError() error {
-	return fmt.Errorf("%w: %w", ErrInvalidOperation, ErrModuloByZero)
-}
